Refatoração aplicando validaçoes de dados

+++ src/Transaction/DeleteTransaction/DeleteTransactionService.ts
@@ -1,13 +1,26 @@
+import { AppError } from "../../Errors/AppError";
 import { ITransactionRepository } from "../../Repository/ITransactionRepository";
 import { TransactionRepository } from "../../Repository/TransactionRepository";
-import { TransactionDTO } from "../dtos/TransactionDTO";
+
+interface IDeleteTransactionServiceProps {
+  id: string;
+}
 
 export class DeleteTransactionService {
   private transactionRepository: ITransactionRepository;
   constructor() {
     this.transactionRepository = new TransactionRepository();
   }
-  async execute(id: string) {
+
+  async execute({ id }: IDeleteTransactionServiceProps) {
+    const transaction = await this.transactionRepository.findOneTransactionById(
+      id
+    );
+
+    if (!transaction) {
+      throw new AppError("Transaction not exists", 400);
+    }
+
     await this.transactionRepository.remove(id);
   }
 }
